+++ src/main/java/wooteco/subway/domain/fare/AgeFare.java
@@ -1,4 +1,4 @@
-package wooteco.subway.domain;
+package wooteco.subway.domain.fare;
 
 import java.util.Arrays;
 import java.util.NoSuchElementException;
