+++ src/test/java/wooteco/subway/domain/AgeFareTest.java
@@ -1,14 +1,11 @@
 package wooteco.subway.domain;
 
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
-import org.springframework.beans.factory.annotation.Value;
+import wooteco.subway.domain.fare.AgeFare;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 class AgeFareTest {
