+++ src/test/java/wooteco/subway/domain/DistanceExtraFareTest.java
@@ -4,6 +4,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
+import wooteco.subway.domain.fare.DistanceExtraFare;
 
 import static org.assertj.core.api.Assertions.assertThat;
